FoundationInternationalization: improve Windows file path handling
This adjusts the Windows paths to account for the handling of file paths
and link resolutions.  There is no `readlink` function, particularly in
the C library.  Instead, we need to open the file, and then query the
I/O Object Manager to determine the final path.  This emulates the
behaviour of `readlink` for Windows.